+++ packages/twenty-server/src/engine/metadata-modules/flat-field-metadata/utils/from-delete-field-input-to-flat-field-metadatas-to-delete.util.ts
@@ -12,13 +12,11 @@ import {
 import { type AllFlatEntityMaps } from 'src/engine/metadata-modules/flat-entity/types/all-flat-entity-maps.type';
 import { findFlatEntityByIdInFlatEntityMaps } from 'src/engine/metadata-modules/flat-entity/utils/find-flat-entity-by-id-in-flat-entity-maps.util';
 import { findFlatEntityByUniversalIdentifierOrThrow } from 'src/engine/metadata-modules/flat-entity/utils/find-flat-entity-by-universal-identifier-or-throw.util';
-import { findManyFlatEntityByIdInFlatEntityMapsOrThrow } from 'src/engine/metadata-modules/flat-entity/utils/find-many-flat-entity-by-id-in-flat-entity-maps-or-throw.util';
 import { findManyFlatEntityByUniversalIdentifierInUniversalFlatEntityMapsOrThrow } from 'src/engine/metadata-modules/flat-entity/utils/find-many-flat-entity-by-universal-identifier-in-universal-flat-entity-maps-or-throw.util';
 import { type FlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type';
 import { computeFlatFieldMetadataRelatedFlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/utils/compute-flat-field-metadata-related-flat-field-metadata.util';
+import { computeSearchFieldMetadataDeletionForDeletedFields } from 'src/engine/metadata-modules/flat-field-metadata/utils/compute-search-field-metadata-deletion-for-deleted-fields.util';
 import { type FlatIndexMetadata } from 'src/engine/metadata-modules/flat-index-metadata/types/flat-index-metadata.type';
-import { recomputeSearchVectorFieldFromSearchFieldMetadatas } from 'src/engine/metadata-modules/flat-object-metadata/utils/recompute-search-vector-field-from-search-field-metadatas.util';
-import { type FlatSearchFieldMetadata } from 'src/engine/metadata-modules/flat-search-field-metadata/types/flat-search-field-metadata.type';
 import { generateFlatIndexMetadataWithNameOrThrow } from 'src/engine/metadata-modules/index-metadata/utils/generate-flat-index.util';
 import { belongsToTwentyStandardApp } from 'src/engine/metadata-modules/utils/belongs-to-twenty-standard-app.util';
 import { type UniversalFlatFieldMetadata } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-field-metadata.type';
@@ -201,89 +199,6 @@ export const fromDeleteFieldInputToFlatFieldMetadatasToDelete = ({
     flatFieldMetadatasToDelete,
     flatIndexesToDelete,
     flatIndexesToUpdate,
-    searchFieldMetadatasToDelete,
-    flatSearchVectorFieldsToUpdate,
-  };
-};
-
-// On field delete, drop each searchFieldMetadata row referencing a deleted field and
-// recompute the affected objects' searchVector (mirrors index removal).
-const computeSearchFieldMetadataDeletionForDeletedFields = ({
-  flatFieldMetadatasToDelete,
-  flatObjectMetadataMaps,
-  flatFieldMetadataMaps,
-  flatSearchFieldMetadataMaps,
-}: {
-  flatFieldMetadatasToDelete: FlatFieldMetadata[];
-} & Pick<
-  AllFlatEntityMaps,
-  | 'flatObjectMetadataMaps'
-  | 'flatFieldMetadataMaps'
-  | 'flatSearchFieldMetadataMaps'
->): {
-  searchFieldMetadatasToDelete: UniversalFlatSearchFieldMetadata[];
-  flatSearchVectorFieldsToUpdate: FlatFieldMetadata<FieldMetadataType.TS_VECTOR>[];
-} => {
-  const deletedFieldIds = new Set(
-    flatFieldMetadatasToDelete.map((flatFieldMetadata) => flatFieldMetadata.id),
-  );
-  const affectedObjectMetadataIds = new Set(
-    flatFieldMetadatasToDelete.map(
-      (flatFieldMetadata) => flatFieldMetadata.objectMetadataId,
-    ),
-  );
-
-  const searchFieldMetadatasToDelete: UniversalFlatSearchFieldMetadata[] = [];
-  const flatSearchVectorFieldsToUpdate: FlatFieldMetadata<FieldMetadataType.TS_VECTOR>[] =
-    [];
-
-  for (const objectMetadataId of affectedObjectMetadataIds) {
-    const flatObjectMetadata = findFlatEntityByIdInFlatEntityMaps({
-      flatEntityId: objectMetadataId,
-      flatEntityMaps: flatObjectMetadataMaps,
-    });
-
-    if (!isDefined(flatObjectMetadata) || !flatObjectMetadata.isSearchable) {
-      continue;
-    }
-
-    const objectSearchFieldMetadatas =
-      findManyFlatEntityByIdInFlatEntityMapsOrThrow<FlatSearchFieldMetadata>({
-        flatEntityMaps: flatSearchFieldMetadataMaps,
-        flatEntityIds: flatObjectMetadata.searchFieldMetadataIds,
-      });
-
-    const rowsToDelete = objectSearchFieldMetadatas.filter(
-      (searchFieldMetadata) =>
-        deletedFieldIds.has(searchFieldMetadata.fieldMetadataId),
-    );
-
-    if (rowsToDelete.length === 0) {
-      continue;
-    }
-
-    searchFieldMetadatasToDelete.push(...rowsToDelete);
-
-    const remainingFieldIds = objectSearchFieldMetadatas
-      .filter(
-        (searchFieldMetadata) =>
-          !deletedFieldIds.has(searchFieldMetadata.fieldMetadataId),
-      )
-      .map((searchFieldMetadata) => searchFieldMetadata.fieldMetadataId);
-
-    const flatSearchVectorFieldToUpdate =
-      recomputeSearchVectorFieldFromSearchFieldMetadatas({
-        flatObjectMetadata,
-        flatFieldMetadataMaps,
-        searchFieldMetadataFieldIds: remainingFieldIds,
-      });
-
-    if (isDefined(flatSearchVectorFieldToUpdate)) {
-      flatSearchVectorFieldsToUpdate.push(flatSearchVectorFieldToUpdate);
-    }
-  }
-
-  return {
     searchFieldMetadatasToDelete,
     flatSearchVectorFieldsToUpdate,
   };
